[feat] 기능 고도화, 불필요한 requeue 제거, reject error 해결, tf 변수 사용법 간소화
