+++ config/test-manager/kustomization.yaml
@@ -5,4 +5,4 @@ kind: Kustomization
 images:
 - name: controller
   newName: 192.168.9.12:5000/tfc-operator
-  newTag: va3
+  newTag: d1112-22
