+++ util/util.go
@@ -1,6 +1,10 @@
 package util
 
 import (
+	"fmt"
+	"strings"
+
+	claimv1alpha1 "github.com/tmax-cloud/tfc-operator/api/v1alpha1"
 	ctrl "sigs.k8s.io/controller-runtime"
 )
 
@@ -21,4 +25,21 @@ func LowestNonZeroResult(i, j ctrl.Result) ctrl.Result {
 	default:
 		return j
 	}
+}
+
+func GetTerraformVariables(tfapply *claimv1alpha1.TFApplyClaim) string {
+
+	if tfapply.Spec.Variable == "" {
+		return ""
+	}
+
+	variableList := strings.Split(tfapply.Spec.Variable, ",")
+
+	cmd := ""
+	for _, v := range variableList {
+		value := strings.Trim(v, " ")
+		cmd += fmt.Sprintf("export TF_VAR_%s; ", value)
+	}
+
+	return cmd
 }
